refactor: Car 내장 함수 테스트 코드 리팩터링
jest.fn()을 사용하지 않아도 테스트할 수 있도록 리팩터링
